% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spare-package.r
\name{add_server}
\alias{add_server}
\title{Add a spare server to a service}
\usage{
add_server(service, ...)
}
\arguments{
\item{service}{Name of the service.}

\item{...}{Server objects to add, or modify. To remove a server,
set it to \code{NULL} here.}
}
\value{
Name of the service, invisibly.
}
\description{
Add a spare server to a service
}

