% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spare-package.r
\name{add_service}
\alias{add_service}
\title{Create a redundant service}
\usage{
add_service(name, ...)
}
\arguments{
\item{name}{Name of the service.}

\item{...}{Server objects, zero or more. Servers can also be added
later, using \code{add_server}. See \code{server}.}
}
\value{
The name of the service, invisibly.
}
\description{
Create a redundant service
}

