% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spare-package.r
\name{remove_service}
\alias{remove_service}
\title{Remove a redundant service}
\usage{
remove_service(name)
}
\arguments{
\item{name}{The name of the service to remove.}
}
\value{
Nothing.
}
\description{
Remove a redundant service
}

