% Generated by roxygen2 (4.0.1): do not edit by hand
\name{server}
\alias{server}
\title{Create a spare server object}
\usage{
server(base_url, priority = 1, timeout = 5, timeout_multiplier = 2)
}
\arguments{
\item{base_url}{Base URL of the server.}

\item{priority}{Non-negative initial odds of using this server, it has
no units. Note that a server with zero odds can be used if all other
servers are down.}

\item{timeout}{Initial timeout for this server, in seconds.}

\item{timeout_multiplier}{The timeout will be multiplied by this value
after an unsuccessful query.}
}
\value{
A spare server object
}
\description{
Create a spare server object
}

