% Generated by roxygen2 (4.0.1): do not edit by hand
\name{service}
\alias{service}
\title{Create a redundant service}
\usage{
service(name, ...)
}
\arguments{
\item{name}{Name of the service.}

\item{...}{Server objects, zero or more. Servers can also be added
later, using \code{add_server}. See \code{server}.}
}
\value{
The name of the service, invisibly.
}
\description{
Create a redundant service
}

