% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spare-package.r
\name{spare_q}
\alias{spare_q}
\title{Make a web query that makes use of redundant servers}
\usage{
spare_q(service, url, fun, ...)
}
\arguments{
\item{service}{Name of the service.}

\item{url}{URL, within the base URL of the server(s). This will be
appended to the base URLs.}

\item{fun}{The function to call. It should have an argument called
\code{url}.}

\item{...}{Additional arguments to pass to \code{fun}.}
}
\value{
The return value of \code{fun}.
}
\description{
Make a web query that makes use of redundant servers
}

