% Generated by roxygen2 (4.0.2): do not edit by hand
\name{server}
\alias{server}
\title{Create a spare server object}
\usage{
server(base_url, priority = 1, timeout = 1)
}
\arguments{
\item{base_url}{Base URL of the server.}

\item{priority}{Non-negative initial odds of using this server, it has
no units. Note that a server with zero odds can be used if all other
servers are down.}

\item{timeout}{Timeout in seconds. After this time, we give up on
the server.}
}
\value{
A spare server object
}
\description{
Create a spare server object
}

